Rename home settings hook
